refactor: break down Header into subcomponents
